pyproject.toml
[project]
name = "metro-pass-breakeven"
version = "2.0.0"
description = "A tool to calculate if a WMATA pass was worth it based on actual usage."
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "python-dotenv>=1.1.0",
]
